requirements.txt
asgiref==3.5.2
autopep8==1.6.0
boto3==1.24.25
botocore==1.27.25
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
ConfigArgParse==1.5.3
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Flask==2.1.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.5
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
locust==2.10.1
MarkupSafe==2.1.1
msgpack==1.0.4
psutil==5.9.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==22.3.0
requests==2.28.1
roundrobin==0.0.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zope.event==4.5.0
zope.interface==5.4.0
